requirements.txt
beautifulsoup==4.9.3
requests==2.25.1
Pillow==8.0.1
jupyter==
pandas==1.1.5
PyDispatcher==2.0.5
